+++ packages/@expo/env/CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### 💡 Others
 
+- Refactor environment loading to remove package state and allow to fetch locally defined environment variables. ([#33633](https://github.com/expo/expo/pull/33633) by [@byCedric](https://github.com/byCedric))
+
 ## 0.4.0 — 2024-10-22
 
 _This version does not introduce any user-facing changes._